third_party/WebKit/LayoutTests/http/tests/inspector/service-workers/service-worker-network-fetch-expected.txt

Tests fetch in Service Workers.

Fetch in worker result: Hello world
http://127.0.0.1:8000/inspector/service-workers/resources/network-fetch-worker.js
resource.type: other
request.failed: false
http://127.0.0.1:8000/inspector/network/resources/resource.php
resource.type: other
request.failed: false
